% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2.R
\docType{data}
\name{climate_anomalies}
\alias{climate_anomalies}
\title{climate_anomalies data set}
\format{
A data frame with 2 columns and 266 rows.
}
\source{
Datahub.io, Climate-change collection
}
\usage{
climate_anomalies
}
\description{
A data frame with different two time series on global mean temperature anomalies (GMTA) and global mean sea level (GMTA).
}
\keyword{datasets}
